chatbot/requirements.txt
google-generativeai>=0.3.0
flask>=2.0.1
python-dotenv>=0.19.0
langchain>=0.1.0
langchain-google-genai>=0.0.5